PMM-7574 move DeleteModal to shared
