+++ public/app/percona/backup/components/StorageLocations/RemoveStorageLocationModal/RemoveStorageLocationModal.tsx
@@ -1,4 +1,4 @@
-import { DeleteModal } from 'app/percona/integrated-alerting/components/DeleteModal';
+import { DeleteModal } from 'app/percona/shared/DeleteModal';
 import React, { FC } from 'react';
 import { WarningBlock } from '../WarningBlock/WarningBlock';
 import { Messages } from './RemoveStorageLocationModal.messages';
