+++ public/app/percona/shared/DeleteModal/DeleteModal.messages.ts
