+++ public/app/percona/shared/DeleteModal/DeleteModal.styles.ts
