+++ public/app/percona/shared/DeleteModal/DeleteModal.tsx
