+++ public/app/percona/shared/DeleteModal/DeleteModal.types.ts
