+++ public/app/percona/shared/DeleteModal/index.ts
